Raywitree

+++ app/Library/DropboxInterface.php
@@ -77,10 +77,16 @@ Class DropboxInterface implements ModelInterface
         return $dbx;
     }
 
-
     public function getAccountInfo()
     {
-        return $this->dbxObj->GetAccountInfo();
+        $info = $this->dbxObj->GetAccountInfo();
+        $nml_info = array(
+            'email' => $info->email,
+            'quota' => $info->quota_info->quota,
+            'used' => $info->quota_info->normal,
+            'remain' => $info->quota_info->quota - $info->quota_info->normal
+        );
+        return (object)$nml_info;
     }
 
     /**
@@ -108,29 +114,45 @@ Class DropboxInterface implements ModelInterface
     // @params $file = String of File Paths on Dropbox
     public function downloadFile($file, $destination = null)
     {
+        if($destination == 'temp'){
+            $destination = $destination . $file;
+            $this->dbxObj->DownloadFile($file,$destination);
+            return $destination;
+        }
+        else{
         $this->dbxObj->DownloadFile($file,$destination);
         header("Content-Type: application/download");
         header("Content-disposition: attachment; filename=". basename($_GET['file']));
         readfile(basename($_GET['file']));
         unlink(basename($_GET['file']));
+        }
     }
     public function uploadFile($file, $destination = null)
     {
-        if (empty($destination)){
-            $destination = $file['name'];
-        } else
-        {
-            $destination = $destination."/".$file['name'];
+        if(is_array($file)){
+            if (empty($destination)){
+                $destination = $file['name'];
+            } else
+            {
+                $destination = $destination."/".$file['name'];
+            }
+
+            $res = $this->dbxObj->UploadFile($file['tmp_name'] ,$destination);
+            return array((object)\GuzzleHttp\json_decode($res));
+        } else {
+            $res = $this->dbxObj->CreateFolder($file);
+            return array((object)\GuzzleHttp\json_decode($res));;
         }
-        return $this->dbxObj->UploadFile($file['tmp_name'] ,$destination)->path;
     }
+
     public function getFiles($file = null)
     {
         return $this->dbxObj->GetFiles($file);
     }
     public function deleteFile($file)
     {
-        return $this->dbxObj->Delete($file);
+        $res = $this->dbxObj->Delete($file);
+        return $res->is_deleted;
     }
 
     public function getLink($file)
@@ -169,7 +191,8 @@ Class DropboxInterface implements ModelInterface
     {
         $lastIndex = strripos($file, "/");
         $new_name = substr($file, 0,$lastIndex+1) . $new_name;
-        return $this->dbxObj->Move($file,$new_name);
+        $res = $this->dbxObj->Move($file,$new_name);
+        return true;
     }
 
     public function getPathName($file)
@@ -179,7 +202,7 @@ Class DropboxInterface implements ModelInterface
 
     public function searchFile($keyword)
     {
-        $list_data =  $this->dbxObj->Search("",$keyword);
+        $list_data =  $this->dbxObj->Search("",urlencode($keyword));
         if(!empty($list_data)){
             $key = array();
             foreach($list_data as $data){
@@ -210,7 +233,7 @@ Class DropboxInterface implements ModelInterface
      * =>(boolean)is_dir
      * =>(string)modified format 'Y m d H:i:s'
      * =>(string)shared
-     * =>(string)token_id
+     * =>(string)provider_logo
      * =>(string)connection_name
      */
     public function normalizeMetaData($list_data, $provider_logo, $connection_name)
@@ -218,16 +241,15 @@ Class DropboxInterface implements ModelInterface
         $format = array();
         foreach ($list_data as $k => $val) {
             $val->is_dir == 1 ? $mime = null : $mime = $val->mime_type;
-            empty($val->shared_folder) ? $sh = false : $sh = true;
             array_push($format,
                 array(
                     'name' => basename($k),
                     'path' => $val->path,
                     'bytes' => $val->bytes,
                     'mime_type' => $mime,
-                    'is_dir' => ($val->is_dir)? true:false, // 1 == Folder, 0 = File
+                    'is_dir' => ($val->is_dir)? true:false,
                     'modified' => date('Y m d H:i:s',strtotime($val->modified)),
-                    'shared' => $sh,
+                    'shared' => null,
                     'provider_logo' => $provider_logo,
                     'connection_name' => $connection_name
                 ));
